Move blobStore related functions out of remotestore.go
